feat: extend qpscs to 8

+++ Source/API/EbSvtAv1Enc.h
@@ -968,12 +968,12 @@ typedef struct EbSvtAv1EncConfiguration {
     bool rtc;
 
     /* @brief compresses the QP hierarchical layer scale to improve temporal video consistency
-    * 0: no compression, original SVT-AV1 scaling
-    * 1-3: enable compression, the higher the number the stronger the compression
-    *      (different frame quality fluctuation/mean quality tradeoffs)
-    * Default is 1 in SVT-AV1-HDR, mainline default is 0.
+    * 0.0: no compression, original SVT-AV1 scaling
+    * 0.0-8.0: enable compression, the higher the number the stronger the compression
+    *         (different frame quality fluctuation/mean quality tradeoffs)
+    * Default is 1.0 in SVT-AV1-HDR, mainline default is 0.0
     */
-    uint8_t qp_scale_compress_strength;
+    double qp_scale_compress_strength;
 
 #if FTR_SFRAME_POSI
     /* @brief Indicates where to insert an S-Frame, only available when sframe_mode is SFRAME_FLEXIBLE_ARF */
